- Move Acter Search Widget file
